Do not decode request path in logs
Log files are openned in raw mode, so UTF-8 characters (>255) cannot be
written. Their bytewise representation should be used instead. But, there is no
good reason to decode paths. Some analyzing tools could be ascii oriented and
this requires an extra encoding to get the original request.

So now, no decoding is done on the request path in access and auth logs.